Modified intiger type to int64

+++ week-2/trunc/trunc.go
@@ -8,5 +8,5 @@ func main() {
 	var input float64
 	fmt.Print("Please provide your floating point value: ")
 	fmt.Scan(&input)
-	fmt.Printf("Truncated number is: %d\n", int32(input))
+	fmt.Printf("Truncated number is: %d\n", int64(input))
 }
